requirements.txt
# Core Dependencies
python-dotenv>=1.0.0

# Supabase Python Client
supabase>=2.0.0

# AI Clients
google-generativeai>=0.3.0
openai>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
tqdm>=4.66.0
